+++ packages/app/drizzle/0000_faulty_archangel.sql
@@ -75,7 +75,7 @@ CREATE TABLE `settings` (
 	`is_biometric_enabled` integer DEFAULT false NOT NULL,
 	`show_cents` integer DEFAULT true NOT NULL,
 	`is_vibration_enabled` integer DEFAULT true NOT NULL,
-    `is_screenshot_protection_enabled` integer DEFAULT true NOT NULL,
+	`is_screenshot_protection_enabled` integer DEFAULT true NOT NULL,
 	FOREIGN KEY (`default_account_id`) REFERENCES `accounts`(`id`) ON UPDATE no action ON DELETE no action,
 	FOREIGN KEY (`default_instrument_id`) REFERENCES `instruments`(`id`) ON UPDATE no action ON DELETE no action
 );
@@ -98,7 +98,6 @@ CREATE TABLE `transactions` (
 	`external_id` text,
 	`operated_at` integer DEFAULT (unixepoch()) NOT NULL,
 	`comment` text DEFAULT '' NOT NULL,
-	`amount` integer DEFAULT 0 NOT NULL,
 	`to_account_id` integer,
 	`from_account_id` integer,
 	`exchange_rate` real NOT NULL,
@@ -115,12 +114,10 @@ CREATE TABLE `transaction_entries` (
 	`type` text NOT NULL,
 	`account_id` integer NOT NULL,
 	`category_id` integer,
-	`instrument_id` integer NOT NULL,
 	`transaction_id` integer NOT NULL,
 	`amount` integer NOT NULL,
 	FOREIGN KEY (`account_id`) REFERENCES `accounts`(`id`) ON UPDATE no action ON DELETE cascade,
 	FOREIGN KEY (`category_id`) REFERENCES `categories`(`id`) ON UPDATE no action ON DELETE cascade,
-	FOREIGN KEY (`instrument_id`) REFERENCES `instruments`(`id`) ON UPDATE no action ON DELETE cascade,
 	FOREIGN KEY (`transaction_id`) REFERENCES `transactions`(`id`) ON UPDATE no action ON DELETE cascade
 );
 --> statement-breakpoint
@@ -215,3 +212,4 @@ VALUES
     (true, 'Clothing & Accessories',     'Shirt',                NULL, false),
     (true, 'Other',                      'CircleDot',            NULL, false),
     (true, 'Emergency Fund',             'AlertTriangle',        NULL, false);
+