+++ packages/app/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.84.1](https://github.com/budgie-at/budgie/compare/v1.84.0...v1.84.1) (2025-12-26)
+
+### Bug Fixes
+
+- create transaction input schema ([d3c5ac0](https://github.com/budgie-at/budgie/commit/d3c5ac081aca3b404d1ce1c62628c526ad09e961))
+- update migrations ([6372666](https://github.com/budgie-at/budgie/commit/63726666e9562534a5bab78fcf9e473721035a97))
+
+# [1.84.0](https://github.com/budgie-at/budgie/compare/v1.83.0...v1.84.0) (2025-12-26)
+
+### Bug Fixes
+
+- change color for amount ([9351ffe](https://github.com/budgie-at/budgie/commit/9351ffeb58703c47a0cd440d30afb7c01e29686a))
+- remove unused instrumentId from transaction entry creation ([ed4f5e1](https://github.com/budgie-at/budgie/commit/ed4f5e13e4bb70c1b4d2c2a61f78a4650b207150))
+- sync translations ([c929f09](https://github.com/budgie-at/budgie/commit/c929f095d0164d74a8d6c330e489586ac86bd1a2))
+- sync translations ([117b5f4](https://github.com/budgie-at/budgie/commit/117b5f4c4d7da663010bd1af5d6999c8c0c36cbf))
+- sync translations ([919d2d4](https://github.com/budgie-at/budgie/commit/919d2d4df11d8aee6a995ac49ee866a706c9946e))
+- update translations ([f5d7151](https://github.com/budgie-at/budgie/commit/f5d715174fafdb322c5947637344df39ff951783))
+
+### Features
+
+- add missing "Unknown" translations for de, es, fr, uk ([4bcfdcf](https://github.com/budgie-at/budgie/commit/4bcfdcf99baf6e3806cb79e49f8a5deb87240495))
+- sync translations ([6793a28](https://github.com/budgie-at/budgie/commit/6793a28cd8b0a815e03de5f5ed27dde09babad57))
+- sync translations ([08a77ac](https://github.com/budgie-at/budgie/commit/08a77ac02c79d8b350d1ae1710a58713e3430bef))
+- sync translations ([9f50471](https://github.com/budgie-at/budgie/commit/9f50471879b5d68973cea44664cad9f592b59c98))
+- sync translations ([279d5ad](https://github.com/budgie-at/budgie/commit/279d5add342a539fa3f47f1d8f6ca9fe4e25e1db))
+- update transaction card ([cf29420](https://github.com/budgie-at/budgie/commit/cf294205c13d0bd69c0b04367eca64c82eeb9443))
+- update transaction card ([64fbb6f](https://github.com/budgie-at/budgie/commit/64fbb6f282496d5e481ba69c9b1680b28a2b1c87))
+- update transaction card ([33c22d7](https://github.com/budgie-at/budgie/commit/33c22d7f7f3cf6d63281b96a2e15f3ac1be7e471))
+
 # [1.83.0](https://github.com/budgie-at/budgie/compare/v1.82.2...v1.83.0) (2025-12-24)
 
 ### Bug Fixes